adding caddy logs

+++ stats/parseLog.py
@@ -3,18 +3,120 @@
 
 import argparse
 import glob
+import ipaddress
 import json
+import logging
 import os
+from urllib.parse import unquote
 
 import pandas as pd
+import sqlalchemy as sqla
 
 # Limit to these postgrest queries
 QUERY_STRINGS = ["enermaps_get_legend"]
-SEL_COLS = ["log_time", "ds_id", "ip", "function", "json_query"]
+
+# Limit to these caddy's URIS
+URIs = [
+    "/enermaps/api/datasets/legend/",
+    "/enermaps/api/db/rpc/enermaps_get_legend?",
+    "/enermaps/api/db/rpc/enermaps_query_geojson?",
+    "/enermaps/api/db/rpc/enermaps_query_table?",
+]
+
+BASE_PATH_PG = "/stats/pg-logs/"
+BASE_PATH_CADDY = "/stats/caddy-logs/"
+
+SEL_COLS = ["timestamp", "ds_id", "country", "function", "json_query", "source"]
+
+REMOVE = True  # remove old logs
+
+DB_HOST = os.environ.get("DB_HOST")
+DB_PORT = os.environ.get("DB_PORT")
+DB_USER = os.environ.get("DB_USER")
+DB_PASSWORD = os.environ.get("DB_PASSWORD")
+DB_DB = os.environ.get("DB_DB")
+
+
+DB_URL = "postgresql://{DB_USER}:{DB_PASSWORD}@{DB_HOST}:{DB_PORT}/{DB_DB}".format(
+    DB_HOST=DB_HOST,
+    DB_PORT=DB_PORT,
+    DB_USER=DB_USER,
+    DB_PASSWORD=DB_PASSWORD,
+    DB_DB=DB_DB,
+)
+
+
+def getCountry(log: pd.DataFrame):
+    """Geolocate ip address."""
+
+    def findCountry(ip, db, information="code"):
+        try:
+            ip = int(ipaddress.ip_address(ip))
+            country = db.loc[(db["start"] < ip) & (db["end"] > ip), information].values[
+                0
+            ]
+        except ValueError:
+            country = "-"
+        return country
+
+    db = pd.read_csv("IP2LOCATION-LITE-DB1/IP2LOCATION-LITE-DB1.CSV", header=None)
+    db.columns = ["start", "end", "code", "country"]
+    log["country"] = ""
+    log["country"] = log["ip"].apply(lambda x: findCountry(x, db))
+
+    return log
+
+
+def safelyJSONdecode(s: str):
+    "Decode json-like string to dict, or keep it as a raw string."
+    try:
+        js = json.loads(s)
+    except (json.decoder.JSONDecodeError, TypeError):
+        js = {"raw_string": r"{}".format(s)}
+    return js
+
+
+def parseCADDYlog(log_file: str):
+    """Parse the original caddy log file."""
+    with open(log_file, "r") as f:
+        dicts = f.read().splitlines()
+    dicts = [json.loads(x) for x in dicts]
+    log = pd.DataFrame.from_records(dicts)
+
+    log["function"] = log["request"].apply(lambda x: x["uri"])
+    log["ip"] = log["request"].apply(
+        lambda x: x.get("headers").get("X-Forwarded-For")[0]
+    )
+    log["timestamp"] = pd.to_datetime(log["ts"], unit="s")
+    log["ds_id"] = -1
+    log["json_query"] = "{}"
+    log.loc[log["function"].str.contains("raster"), "ds_id"] = (
+        log.loc[log["function"].str.contains("raster"), "function"]
+        .str.split("/")
+        .str[-2]
+    )
+    log.loc[log["function"].str.contains("?parameters", regex=False), "json_query"] = (
+        log.loc[log["function"].str.contains("?parameters", regex=False), "function"]
+        .str.split("parameters=")
+        .str[-1]
+    )
+    log["json_query"] = log["json_query"].apply(lambda x: unquote(x))
+    log["json_query"] = log["json_query"].str.replace("+", "", regex=False)
+    log["json_query"] = log["json_query"].apply(lambda x: safelyJSONdecode(x))
+
+    log["time"] = pd.to_datetime(log["ts"], unit="s")
+
+    selected = []
+    for select in URIs:
+        selected.append(log.loc[log["function"].str.startswith(select), :])
+
+    parsed_log = pd.concat(selected)
+    parsed_log["source"] = "caddy"
+    return parsed_log
 
 
 def parsePGlog(log_file: str):
-    """Parse the original log file."""
+    """Parse the original PG log file."""
     header = [
         "log_time",
         "user_name",
@@ -42,23 +144,13 @@ def parsePGlog(log_file: str):
         "backend_type",
     ]
     try:
-        log = pd.read_csv(
-            log_file, header=None, error_bad_lines=False, low_memory=False
-        )
+        log = pd.read_csv(log_file, header=None, on_bad_lines="skip", low_memory=False)
         log.columns = header
     except (pd.errors.EmptyDataError, pd.errors.ParserError):
         print("Cannot decode the content of the log file")
         log = pd.DataFrame()
 
     parsed_log = []
-
-    def safelyJSONdecode(s: str):
-        "Decode json-like string to dict, or keep it as a raw string."
-        try:
-            js = json.loads(s)
-        except (json.decoder.JSONDecodeError, TypeError):
-            js = {"raw_string": r"{}".format(s)}
-        return js
 
     if log.shape[0] > 0:
         print("Parsing")
@@ -68,10 +160,11 @@ def parsePGlog(log_file: str):
             log["message"]
             .str.lower()
             .str.startswith("statement: select * from enermaps"),
-            ["log_time", "message"],
+            :,
         ]
+
         if queries.shape[0] > 0:
-            queries["message"] = log["message"].str.replace("\n", "")
+            queries["message"] = queries["message"].str.replace("\n", "")
             # Get the query
             queries["json_query"] = queries.message.str.extract(r"(?<=\(')(.*)(?='\))")
             queries["json_query"] = queries["json_query"].apply(
@@ -95,16 +188,17 @@ def parsePGlog(log_file: str):
         # Parse log
         # PostgREST queries
         queries = log.loc[
-            ~log["detail"].isnull(),  # the parameters are saved in the detail field
-            ["log_time", "message", "detail", "session_id", "session_line_num"],
+            (~log["detail"].isnull())
+            & (
+                log["message"].str.startswith("execute ")
+            ),  # the parameters are saved in the detail field
+            :,
         ]
         selected = []
         for string in QUERY_STRINGS:
             selected.append(queries.loc[queries.message.str.contains(string), :])
         queries = pd.concat(selected)
-
         if queries.shape[0] > 0:
-            print("PostgREST queries")
             queries["detail"] = queries["detail"].str.replace("\n", "")
             # Get the query
             queries["json_query"] = queries["detail"].str.extract(
@@ -116,7 +210,10 @@ def parsePGlog(log_file: str):
             # Get the dataset_id
             ds_id = []
             for d in queries.json_query:
-                parameters = d.get("parameters", {})
+                if "id" in d.keys():
+                    parameters = d
+                else:
+                    parameters = d.get("parameters", {})
                 if isinstance(parameters, str):
                     parameters = json.loads(parameters)
                 if "data.ds_id" in parameters.keys():
@@ -127,14 +224,15 @@ def parsePGlog(log_file: str):
                     ds_id.append(-1)
             queries["ds_id"] = ds_id
             # Drop duplicate rows
-            queries = queries.groupby("session_id").first()
+            queries = queries.groupby("log_time").first().reset_index()
 
             # Find PostGrest function
             queries["function"] = queries["message"].str.extract(
                 r'(?<=SELECT "public".")(.*)(?="\()'
             )
+
             # Find geolocalization info based on session_id
-            geolocal = log.loc[log["session_id"].isin(queries.index), :]
+            geolocal = log.loc[log["session_id"].isin(queries["session_id"]), :]
             # Get the IP address
             geolocal = geolocal.loc[
                 geolocal["message"].str.contains("request.header.x-forwarded-for"), :
@@ -147,7 +245,10 @@ def parsePGlog(log_file: str):
             # Merge geolocalization info with query info
             if geolocal.shape[0] > 0:
                 queries = pd.merge(
-                    geolocal[["ip", "session_id"]], queries, on="session_id"
+                    geolocal[["ip", "session_id"]],
+                    queries,
+                    on="session_id",
+                    how="outer",
                 )
             else:
                 queries["ip"] = None
@@ -160,15 +261,16 @@ def parsePGlog(log_file: str):
         if len(parsed_log) > 0:
             parsed_log = pd.concat(parsed_log, ignore_index=True)
             parsed_log["log_time"] = pd.to_datetime(parsed_log["log_time"])
-            parsed_log = parsed_log.sort_values("log_time", ascending=True)
+            parsed_log["timestamp"] = pd.to_datetime(parsed_log["session_start_time"])
+            parsed_log = parsed_log.sort_values("session_start_time", ascending=True)
 
             # Drop duplicates
             parsed_log = (
-                parsed_log.groupby(["log_time", "ip", "function"], dropna=False)
+                parsed_log.groupby(["session_start_time"], dropna=False)
                 .first()
                 .reset_index()
             )
-
+            parsed_log["source"] = "pg"
             return parsed_log
 
 
@@ -182,27 +284,61 @@ def saveCSV(parsed_log: pd.DataFrame, parsed_log_file: str, sel_cols: list = SEL
 if __name__ == "__main__":
     print("calling parsing")
     parser = argparse.ArgumentParser(description="Create log")
-    parser.add_argument("source_log_file", default="all")
+    parser.add_argument("--source_log_file", default="all", required=False)
     parser.add_argument("--parsed_log_file", "-o", default="tmp.csv", required=False)
+    parser.add_argument("--sql", action="store_true")
 
     args = parser.parse_args()
 
     if args.source_log_file == "all":
+        pglogs = []
+        caddylogs = []
         # By default parse all the log files but the last one and remove them
         for log_file in sorted(
-            glob.glob("/db-data/pg_log/*.csv"), key=os.path.getmtime
+            glob.glob("{}*.csv".format(BASE_PATH_PG)), key=os.path.getmtime
         )[:-1]:
             print("Reading {}".format(log_file))
             pglog = parsePGlog(log_file)
-            saveCSV(pglog, os.path.abspath(os.path.join("stats", "pg_log.csv")))
-            # Remove source log files
+            pglogs.append(pglog)
+        if REMOVE:
+            logging.info("Remove source pglog files")
             os.remove(log_file)
             log_file2 = log_file.replace(".csv", "")
             if os.path.exists(log_file2):
                 os.remove(log_file2)
+
+        for log_file in sorted(
+            glob.glob("{}*.log".format(BASE_PATH_CADDY)), key=os.path.getmtime
+        )[:-1]:
+            print("Reading {}".format(log_file))
+            caddylog = parseCADDYlog(log_file)
+            caddylogs.append(caddylog)
+            if REMOVE:
+                logging.info("Remove source caddy log files")
+                os.remove(log_file)
+        logs = pd.concat([*pglogs, *caddylogs], ignore_index=True)
+        pglogs = pd.concat(pglogs, ignore_index=True)
+        caddylogs = pd.concat(caddylogs, ignore_index=True)
+        # Geolocalize logs
+        logs = getCountry(logs)
+        # Prepare df
+        logs = logs.loc[:, SEL_COLS]
+        logs["json_query"] = logs["json_query"].apply(lambda x: json.dumps(x))
+        if args.sql:
+            db_engine = sqla.create_engine(DB_URL)
+            logging.info("Loading to PostgreSQL...")
+            logs.to_sql("stats", db_engine, if_exists="append", index=False)
+            logging.info("Done.")
+        else:
+            saveCSV(pglog, os.path.abspath(os.path.join("stats", "pg_log.csv")))
     else:
         print("Manually loading log file")
-        pglog = parsePGlog(
-            os.path.abspath(os.path.join("db-data", "pg_log", args.source_log_file))
-        )
-        saveCSV(pglog, os.path.abspath(os.path.join("stats", args.parsed_log_file)))
+        if "pg" in args.source_log_file:
+            log = parsePGlog(
+                os.path.abspath(os.path.join(BASE_PATH_PG, args.source_log_file))
+            )
+        elif "caddy" in args.source_log_file:
+            log = parseCADDYlog(
+                os.path.abspath(os.path.join(BASE_PATH_CADDY, args.source_log_file))
+            )
+        saveCSV(log, os.path.abspath(os.path.join(args.parsed_log_file)))
